requirements.txt
python-chess==0.24.1
numpy==1.15.4
pandas==0.25.3
seaborn==0.9.0
